Update needed packages and GitHub Actions workflow
Both the setup script and Github Actions workflows have been updated. The 'youtube-dl' package was replaced with 'yt-dlp' and 'mutagen' was updated to a newer version in setup.py. Meanwhile, the GitHub Actions workflows were significantly revised to support newer Python versions and to improve the publishing strategy to PyPI.